Modification to Error Output on Line 84
In the event that a matrix would have a finite number of solutions but would require row swapping (which isn't allowed for base LU factorization), it would originally display that the System had infinitely many solutions. I modified the output message to clearly specify the two cases that will trigger during the check. Either
1) The System has Infinitely many solutions.
2) The System requires row swapping to be solved via the program.
Otherwise, the code should be basically finished.